requirements.txt
Django==2.2.5
Pillow==8.1.0
pytz==2020.5
shell==1.0.1
sqlparse==0.4.1
